requirements.txt
bleak==1.1.1
bluez-peripheral==0.1.7
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dbus-fast==2.44.5
dbus-next==0.2.3
idna==3.11
paho-mqtt==2.1.0
pycparser==2.23
requests==2.32.5
urllib3==2.5.0
